+++ src/Components/DashBoardAdmin/OwnerInfo/OwnerInfo.module.css
@@ -1,0 +1,94 @@
+.form {
+	margin-top: 20px;
+	border-radius: 15px;
+	padding: 20px;
+	height: 100%;
+}
+
+.form > div {
+	margin-bottom: 20px;
+}
+
+.form label {
+	display: block;
+	font-weight: 600;
+	font-size: 18px;
+	margin-bottom: 5px;
+}
+
+.form input,
+.form textarea {
+	width: 100%;
+    background-color: #f4f4f4;
+	border-radius: 15px;
+    border: none;
+	font-size: 18px;
+	padding: 5px 0px 5px 10px;
+	margin-bottom: 5px;
+	color: #2E2D2D;
+}
+
+.form input:focus {
+	outline: none;
+	border: 2px solid #ffa726;
+}
+
+.error {
+	color: #e92b2d;
+	font-weight: 600;
+	font-size: 12px;
+}
+
+.form button[type="submit"] {
+	display: block;
+	background: #ffa726;
+	font-weight: 600;
+	font-family: 'Open Sans', sans-serif;
+	border: none;
+	cursor: pointer;
+	width: 100%;
+	padding: 10px;
+	border-radius: 15px;
+	color: #fff;
+	font-size: 16px;
+	transition: .3s ease all;
+}
+
+.form button[type="submit"]:hover {
+	background-color: #ffa319;
+    border-color: #ffffff;
+    box-shadow: 2px 2px 10px rgba(128, 128, 128, 0.420);
+}
+
+.success {
+	text-align: center;
+	font-size: 14px;
+	padding-top: 20px;
+	color: #02a802;
+}
+
+.form input[type="radio"] {
+	width: auto;
+}
+
+.form select {
+    width: 100%;
+    background-color: #f4f4f4;
+    border-radius: 15px;
+    border: none;
+    font-size: 18px;
+    padding: 5px 0px 5px 10px;
+    margin-bottom: 5px;
+    color: #2E2D2D;
+}
+
+.form select:focus {
+    outline: none;
+    border: 2px solid #ffa726;
+}
+
+.form select option {
+    background-color: #f4f4f4;
+    color: #2E2D2D;
+    scrollbar-color: #ffa726;
+}
